refactor(monitoring): remove monitoring lapper and replace
