+++ src/main/java/telraam/database/daos/DetectionDAO.java
@@ -9,6 +9,7 @@ import org.jdbi.v3.sqlobject.statement.SqlQuery;
 import org.jdbi.v3.sqlobject.statement.SqlUpdate;
 import telraam.database.models.Detection;
 
+import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
 
@@ -54,6 +55,11 @@ public interface DetectionDAO extends DAO<Detection> {
     @SqlQuery("SELECT * FROM detection WHERE id > :id ORDER BY id LIMIT :limit")
     @RegisterBeanMapper(Detection.class)
     List<Detection> getSinceId(@Bind("id") int id, @Bind("limit") int limit);
+
+
+    @SqlQuery("SELECT * FROM detection WHERE baton_id = :batonId AND timestamp > :timestamp ORDER BY timestamp DESC LIMIT 1")
+    @RegisterBeanMapper(Detection.class)
+    Optional<Detection> latestDetectionByBatonId(@Bind("batonId") int batonId, @Bind("timestamp") Timestamp timestamp);
 
     @SqlQuery("""
             WITH bso AS (SELECT teamid, newbatonid, timestamp AS current_timestamp, LEAD(timestamp) OVER (PARTITION BY teamid ORDER BY timestamp) next_baton_switch FROM batonswitchover)
